fixing issues in cloud

+++ R/dataSetup.R
@@ -320,11 +320,12 @@ bycatchSetup <- function(
     }
 
   #Clean up: delete the figures/ directory after rendering
-  fig_dir <- file.path(dirname(mkd), "figures")
-  if (dir.exists(fig_dir)) {
-    unlink(fig_dir, recursive = TRUE)
-
+  if(!is.null(mkd)) {
+    fig_dir <- file.path(dirname(mkd), "figures")
+    if (dir.exists(fig_dir)) {
+      unlink(fig_dir, recursive = TRUE)
     }
+  }
 
   return(output)
